requirements.txt
python-dotenv
mysql-connector-python
pymysql
sqlalchemy
